Move tracing to respective agent, controller and common directories
