+++ common/lib/log_utils.py
@@ -65,7 +65,7 @@ def configure_logging(
 
     # quieten the trace logs if running locally
     trace_log_level = os.environ.get("TRACE_LOG_LEVEL", "ERROR")
-    logging.getLogger("jobrunner.tracing").setLevel(trace_log_level)
+    logging.getLogger("controller.tracing").setLevel(trace_log_level)
 
     # We attach a custom handler for uncaught exceptions to display error
     # output from failed subprocesses
