+++ controller/create_or_update_jobs.py
@@ -20,7 +20,7 @@ from common.lib.github_validators import (
     validate_branch_and_commit,
     validate_repo_url,
 )
-from controller import config
+from controller import config, tracing
 from controller.actions import get_action_specification
 from controller.lib.database import exists_where, insert, transaction, update_where
 from controller.models import Job, SavedJobRequest, State, StatusCode
@@ -29,7 +29,6 @@ from controller.reusable_actions import (
     ReusableActionError,
     resolve_reusable_action_references,
 )
-from jobrunner import tracing
 
 
 log = logging.getLogger(__name__)
